+++ crates/kreuzcrawl-wasm/src/lib.rs
@@ -4399,6 +4399,7 @@ impl From<WasmScrapeResult> for kreuzcrawl::ScrapeResult {
             extraction_meta: val.extraction_meta.map(Into::into),
             screenshot: Default::default(),
             downloaded_document: val.downloaded_document.map(Into::into),
+            browser: None,
         }
     }
 }